add a new city to the remote state
Add the lat and lng of the selected city on the map to the URL as
states. The Form component gets lat and lng from the URL and sends a
GET request to an API to get an object containing the data about the
selected city on the map. Send a POST request to the json server by
clicking on the ADD button to update the remote state (data.json file).
Update the context so that the UI state gets sync with the remote state.